Clean up unused scaffolding.
